split test_commands to have a separate module for testing CommandMeta; finished organizing tests into packages
