+++ tests/test_commands/test_param_registry.py
